Change widget structure
