+++ src/package.xml
@@ -2,10 +2,10 @@
 <package xmlns="http://www.mendix.com/package/1.0/">
     <clientModule name="PullToRefresh" version="0.1.0" xmlns="http://www.mendix.com/clientModule/1.0/">
         <widgetFiles>
-            <widgetFile path="com/mendix/widget/PullToRefresh/PullToRefresh.xml"/>
+            <widgetFile path="PullToRefresh/widget/PullToRefresh.xml"/>
         </widgetFiles>
         <files>
-            <file path="com/mendix/widget/PullToRefresh"/>
+            <file path="PullToRefresh/widget"/>
         </files>
     </clientModule>
 </package>
